refactor: accept feedback
