+++ whenthen-web/src/pages/MockPage/index.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import useRootData from '../../hooks/useRootData';
 import MockComponent from '../../components/MockComponent';
-import TimelineCardComponent from '../../components/TimelineCardComponent';
 import stylesDesktopDefault from './DesktopDefault.module.scss';
 import stylesMobileDefault from './MobileDefault.module.scss';
 
@@ -13,8 +12,7 @@ const MockPage = () => {
 
   const styles = isDesktop ? stylesDesktopDefault : stylesMobileDefault;
 
-  return <TimelineCardComponent />;
-  //return <MockComponent message={'hello, world'} highlight={true} />;
+  return <MockComponent message={'hello, world'} highlight={true} />;
 };
 
 export default MockPage;
